# FIX-#4818, PERF-#4825: Fix where by using the new n-ary operator