--- a/modin/core/dataframe/algebra/binary.py
+++ b/modin/core/dataframe/algebra/binary.py
@@ -95,7 +95,7 @@ class Binary(Operator):
                     )
                 else:
                     return query_compiler.__constructor__(
-                        query_compiler._modin_frame.nary_op(
+                        query_compiler._modin_frame.n_ary_op(
                             lambda x, y: func(x, y, *args, **kwargs),
                             [other._modin_frame],
                             join_type=join_type,