--- a/modin/core/dataframe/pandas/partitioning/partition_manager.py
+++ b/modin/core/dataframe/pandas/partitioning/partition_manager.py
@@ -1263,9 +1263,14 @@ class PandasDataframePartitionManager(ClassLogger, ABC):
 
     @classmethod
     @wait_computations_if_benchmark_mode
-    def binary_operation(cls, left, func, right):
-        """
-        Apply a function that requires two ``PandasDataframe`` objects.
+    def nary_operation(cls, left, func, right: list):
+        r"""
+        Apply a function that requires multiple ``PandasDataframe`` objects.
+
+        This method assumes that all the partitions of the dataframes in left
+        and right have the same dimensions. For each position i, j in each
+        dataframe's partitions, the result has a partition at (i, j) whose data
+        is func(let_partitions[i,j], \*each_right_partitions[i,j]).
 
         Parameters
         ----------
@@ -1273,8 +1278,8 @@ class PandasDataframePartitionManager(ClassLogger, ABC):
             The partitions of left ``PandasDataframe``.
         func : callable
             The function to apply.
-        right : np.ndarray
-            The partitions of right ``PandasDataframe``.
+        right : list of np.ndarray
+            The list of partitions of other ``PandasDataframe``.
 
         Returns
         -------
@@ -1283,7 +1288,7 @@ class PandasDataframePartitionManager(ClassLogger, ABC):
         """
         func = cls.preprocess_func(func)
 
-        def get_right_block(row_idx, col_idx):
+        def get_right_block(right_partitions, row_idx, col_idx):
             blocks = right[row_idx][col_idx].list_of_blocks
             # TODO Resolve this assertion as a part of #4691, because the current implementation assumes
             # that partition contains only 1 block.
@@ -1297,7 +1302,10 @@ class PandasDataframePartitionManager(ClassLogger, ABC):
                 [
                     part.apply(
                         func,
-                        get_right_block(row_idx, col_idx),
+                        *(
+                            get_right_block(right_partitions, row_idx, col_idx)
+                            for right_partitions in right
+                        ),
                     )
                     for col_idx, part in enumerate(left[row_idx])
                 ]