--- a/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition_manager.py
+++ b/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition_manager.py
@@ -131,16 +131,13 @@ class PandasOnRayDataframePartitionManager(GenericRayDataframePartitionManager):
         ray.wait(blocks, num_returns=len(blocks))
 
 
-<<<<<<< HEAD
 def _make_wrapped_method(name: str):
     """
     Define new attribute that should work with progress bar.
-
     Parameters
     ----------
     name : str
         Name of `GenericRayDataframePartitionManager` attribute that should be reused.
-
     Notes
     -----
     - `classmethod` decorator shouldn't be applied twice, so we refer to `__func__` attribute.