## What do these changes do?

- [x] commit message follows format outlined [here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #4818, #4825
- [x] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->
- [x] added (Issue Number: PR title (PR Number)) and github username to release notes for next major release <!-- e.g. DOCS-#4077: Add release notes template to docs folder (#4078) -->

**Performance Results**

For three dataframes with same row and column labels:

<details>

<summary>Benchmark Script (ipython)</summary>

```python
import modin.pandas as pd
import numpy as np
from modin.config import BenchmarkMode

BenchmarkMode.put(True)
df = pd.DataFrame(np.random.randint(0,100,(2**20,2**8)))
cond = df % 2 == 0
other = pd.DataFrame(np.random.randint(0,100,(2**20,2**8)))

%time result = df.where(cond, other)
```
</details>

Wall times (master at f7fd5595429e4edbd147af5c514312339f421813): 15.7 s, 15 s
new implementation times (this branch at c05ef5e56cf5547115aa08870cf89d4a8a47c6f2): 1.59 sec, 1.58 sec
Wall times (pandas 1.4.3): 6.43 s, 6.57 s

For three dataframes each with different row and column -labels:

<details>

<summary>Benchmark Script (ipython)</summary>

```python
import modin.pandas as pd
import numpy as np
import time
import ray

df = pd.DataFrame(np.random.randint(0,100,(2**20,2**8)))
cond = pd.DataFrame(np.random.choice(a=[False, True], size=(2**20, 2**8)), index=df.index[::-1], columns=df.columns[::-1])
other = pd.DataFrame(np.random.randint(0,100,(2**20,2**8)), index=df.index[-1:].append(df.index[:-1]), columns=df.columns[-1:].append(df.columns[:-1]))

# Benchmark mode artificially slows modin down because we have to block on
# each reindex. Instead, manually wait for the where to finish.
def get_timed_where_result():
  start = time.time()
  result = df.where(cond, other)
  all_futures = [x._data for x in result._query_compiler._modin_frame._partitions.flatten()]
  ray.wait(all_futures, num_returns=len(all_futures))
  print(f'where took {time.time() - start}')
  return result

result = get_timed_where_result()
```
</details>

master at f7fd5595429e4edbd147af5c514312339f421813 gives the wrong result.
times (this branch at c05ef5e56cf5547115aa08870cf89d4a8a47c6f2): 6.43 sec, 6.37 sec
times (pandas 1.4.3): 7.16 sec, 7.11 sec

**System information (Amazon EC2)**:

- AMI name: ubuntu/images/hvm-ssd/ubuntu-focal-20.04-amd64-server-20211129
- instance type: t2.2xlarge
- Ubuntu 20.04.3 LTS 
- 8 physical CPUs
- Memory: 32 GiB
- Python: 3.9.7