@pyrito 2022-08-16
@mvashishtha that's a pretty large speedup! How does pandas do?