@mvashishtha 2022-08-16
> that's a pretty large speedup! How does pandas do?

@pyrito I added wall times for pandas. It looks like on this particular benchmark Modin is now about 3x the performance of pandas.