@mvashishtha 2022-08-16
> I wonder if we should name the PR PERF- or FEAT-, rather than FIX-, though.

@vnlitvinov to be more precise, I created #4825 and referenced both the original issue and that one in the new commit description.