@vnlitvinov 2022-08-19
If we're discarding older binary operators in favour of new N-ary, please rename all relevant `binary.py`-like named files and `Binary()`-like named classes.